% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hospital_metadata.R
\name{get_hospital_metadata}
\alias{get_hospital_metadata}
\title{Get metadata on all hospitals covered by the "tempos" API}
\usage{
get_hospital_metadata(output_format = c("json", "data_frame"),
  request_headers = "")
}
\arguments{
\item{output_format}{Character which defines the format of the final output resulting form the API call. Two options:
(1) \code{output_format = "data_frame"} returns a tibble object;
(2) \code{output_format = "json"} returns the data as a json file}

\item{request_headers}{named character vector with the HTTP headers to be added to \code{\link[httr]{GET}} via \code{\link[httr]{add_headers}}. Defaults to \code{NULL}}
}
\value{
json string or tibble containing the relevant metadata.
}
\description{
This function provides a wrapper to the 'api.php/institution' endpoint from the following domain: \link{http://tempos.min-saude.pt}.
Crucial for extracting the wait times, see \code{\link{get_wait_times}} has it contains the "hospital id"s as well as logical variables for the data available per website
}
\examples{
\dontrun{
library(esperaR)
library(jsonlite)

## pull the metadata as a json file
hospital_meta_json <- get_hospital_metadata(output_format = "json")

prettify(hospital_meta_json)

## Adding a user-agent header to the get request and returning a tibble
hospital_meta_df <- get_hospital_metadata(output_format = "data_frame", request_headers = list("User-Agent" = "Mozilla/5.0 (X11; Linux x86_64; rv:74.0) Gecko/20100101 Firefox/74.0"))

}
}
