% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wait_times.R
\name{get_wait_times}
\alias{get_wait_times}
\title{Get wait times for emergencies, consultations, and surgeries in the Hospitals covered by the "tempos" API}
\usage{
get_wait_times(hospital_id = NULL, output_format = c("json",
  "data_frame"), request_headers = NULL, data_type = c("emergency",
  "consultation", "surgery"))
}
\arguments{
\item{hospital_id}{Integer containing the hospital id, get it through \code{\link{get_hospital_metadata}} under the collumn \code{id}}

\item{output_format}{defines the format of the final output resulting form the API call. Two options:
(1) \code{output_format = "data_frame"} returns a tibble object;
(2) \code{output_format = "json"} returns the data as a json file}

\item{request_headers}{character vector with the HTTP headers to be added to \code{\link[httr]{GET}} via \code{\link[httr]{add_headers()}}. Defaults to \code{NULL}}

\item{data_type}{Character string determining the type of data to request. Can be either (1) \code{"emergency"}, (2) \code{"consultation"}, or \code{"surgery"}}
}
\value{
json string or tibble containing the relevant metadata. The relevant wait times for \code{"emergency"} are in both seconds as integer or \code{\link[lubridate]{ymd_hms}}. For \code{"consultation"} or \code{"surgery"} they unit of analysis is day as integer. All datasets also contain a variable measuring wait times by the number of people waiting.
}
\description{
This function provides a wrapper to the API endpoints for getting the live wait times for emergencies, consultations, and surgeries at the following domain: [http://tempos.min-saude.pt]http://tempos.min-saude.pt.
\code{\link{get_wait_times_all}} runs this function on all existing ids.
}
\examples{
library(jsonlite)
library(esperaR)

\dontrun{
library(ggplot2)
## get the metadata for extracting the ids
metadata <- get_hospital_metadata(output_format = "data_frame") \%>\%
 sample_n(., 1)

## extract emergency data
wait_dta <- get_wait_times(hospital_id = metadata$id,
                          output_format = "data_frame",
                          request_headers = c(from = "fooh@baah.pt"),
                          data_type = "emergency")

wait_dta \%>\%
 ggplot(aes(triage_type, wait_time_secs)) +
 geom_boxplot()
}
}
