% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital_data_availability.R
\name{hospital_data_availability}
\alias{hospital_data_availability}
\title{Check data availability for a Hospital covered by the "tempos" API}
\usage{
hospital_data_availability(hospital_id = NULL,
  hospital_metadata = NULL)
}
\arguments{
\item{hospital_id}{Integer containing the hospital id, get it through \code{\link{get_hospital_metadata}} under the collumn \code{id}}

\item{hospital_metadata}{a "data.frame" or NULL (default). If NULL, the function makes an API call to extract the hospitals metadata
using \code{\link{get_hospital_metadata}}.}
}
\value{
Returns a tibble containing three variables: "data_type", character referring to the type of data provided by the endpoint, "availability",
logical values providing information on the data availability in that hospital
}
\description{
This function takes a hospital id and returns the data availability in said hospital. Makes use of the hospital metadata, either by resorting to the wrapper \code{\link{get_hospital_metadata}} or when supplied by the user,
}
\examples{
my_metadata <- get_hospital_metadata(output_format = "data_frame", request_headers = "")

hospital_data_availability(hospital_id = sample(my_metadata$id, 1), hospital_metada = my_metadata)

}
