% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{prep_id}
\alias{prep_id}
\title{Error handling helper functions
A couple of error handling helper fuctions used in \code{\link{get_hospital_metadata}}, \code{\link{get_wait_times}} and \code{\link{hospital_data_availability}}}
\usage{
prep_id(hospital_id = NULL)
}
\description{
Error handling helper functions
A couple of error handling helper fuctions used in \code{\link{get_hospital_metadata}}, \code{\link{get_wait_times}} and \code{\link{hospital_data_availability}}
}
